pyproject.toml
[tool.poetry]
name = "test_service"
version = "0.1.0"
description = ""
authors = ["Adamozo <naworsk.adam@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.3"
pika = "^1.3.2"
uvicorn = "^0.29.0"
SQLAlchemy = "^2.0.29"
yapf = "^0.40.2"
aio-pika = "^9.4.1"
dataclasses-json = "^0.6.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
websockets = "^12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
split_before_logical_operator = true
column_limit = 110
blank_line_before_nested_class_or_def = true
blank_lines_around_top_level_definition = 2